+++ examples/desktop/2-capture_grab.cpp
@@ -73,8 +73,8 @@ int main(int argc, char** argv) {
 
     /// 3000 ms timeout when grabbing frames
     while (auto frame = cameraProvider.grab(3000)) {
-        printf("VideoFrame %lld grabbed: width = %d, height = %d, bytes: %d\n", frame->frameIndex, frame->width, frame->height,
-               frame->sizeInBytes);
+        printf("VideoFrame %d grabbed: width = %d, height = %d, bytes: %d\n", (int)frame->frameIndex, frame->width, frame->height,
+               (int)frame->sizeInBytes);
         if (auto dumpFile = ccap::dumpFrameToDirectory(frame.get(), captureDir); !dumpFile.empty()) {
             std::cout << "VideoFrame saved to: " << dumpFile << std::endl;
         } else {
